+++ pcs_test/tier1/cluster/test_config_show.py
@@ -12,89 +12,96 @@ from pcs_test.tools.pcs_runner import PcsRunner
 from .common import fixture_corosync_conf_minimal
 
 
+def fixture_text_output(no_cluster_uuid=False):
+    text_output = "Cluster Name: cluster_name\n"
+    if not no_cluster_uuid:
+        text_output += "Cluster UUID: cluster_uuid\n"
+    text_output += dedent(
+        """\
+        Transport: knet
+        Nodes:
+          node1:
+            Link 0 address: node1_addr
+            nodeid: 1
+          node2:
+            Link 0 address: node2_addr
+            nodeid: 2
+        Transport Options:
+          ip_version: ipv6
+        Crypto Options:
+          cipher: aes256
+          hash: sha256
+      """
+    )
+    return text_output
+
+
+def fixture_json_output(no_cluster_uuid=False):
+    return (
+        json.dumps(
+            {
+                "cluster_name": "cluster_name",
+                "cluster_uuid": None if no_cluster_uuid else "cluster_uuid",
+                "transport": "KNET",
+                "totem_options": {},
+                "transport_options": {"ip_version": "ipv6"},
+                "compression_options": {},
+                "crypto_options": {"cipher": "aes256", "hash": "sha256"},
+                "nodes": [
+                    {
+                        "name": "node1",
+                        "nodeid": "1",
+                        "addrs": [
+                            {
+                                "addr": "node1_addr",
+                                "link": "0",
+                                "type": "FQDN",
+                            }
+                        ],
+                    },
+                    {
+                        "name": "node2",
+                        "nodeid": "2",
+                        "addrs": [
+                            {
+                                "addr": "node2_addr",
+                                "link": "0",
+                                "type": "FQDN",
+                            }
+                        ],
+                    },
+                ],
+                "links_options": {},
+                "quorum_options": {},
+                "quorum_device": None,
+            }
+        )
+        + "\n"
+    )
+
+
+def fixture_cmd_output(no_cluster_uuid=False):
+    cmd_output = dedent(
+        """\
+        pcs cluster setup cluster_name \\
+          node1 addr=node1_addr \\
+          node2 addr=node2_addr \\
+          transport \\
+          knet \\
+              ip_version=ipv6 \\
+            crypto \\
+              cipher=aes256 \\
+              hash=sha256"""
+    )
+    if no_cluster_uuid:
+        cmd_output += " \\\n  --no-cluster-uuid"
+
+    return cmd_output + "\n"
+
+
 class ClusterConfigMixin(AssertPcsMixin):
     # pylint: disable=invalid-name
     command = None
-
-    @staticmethod
-    def fixture_text_output():
-        return dedent(
-            """\
-            Cluster Name: cluster_name
-            Transport: knet
-            Nodes:
-              node1:
-                Link 0 address: node1_addr
-                nodeid: 1
-              node2:
-                Link 0 address: node2_addr
-                nodeid: 2
-            Transport Options:
-              ip_version: ipv6
-            Crypto Options:
-              cipher: aes256
-              hash: sha256
-          """
-        )
-
-    @staticmethod
-    def fixture_json_output():
-        return (
-            json.dumps(
-                {
-                    "cluster_name": "cluster_name",
-                    "transport": "KNET",
-                    "totem_options": {},
-                    "transport_options": {"ip_version": "ipv6"},
-                    "compression_options": {},
-                    "crypto_options": {"cipher": "aes256", "hash": "sha256"},
-                    "nodes": [
-                        {
-                            "name": "node1",
-                            "nodeid": "1",
-                            "addrs": [
-                                {
-                                    "addr": "node1_addr",
-                                    "link": "0",
-                                    "type": "FQDN",
-                                }
-                            ],
-                        },
-                        {
-                            "name": "node2",
-                            "nodeid": "2",
-                            "addrs": [
-                                {
-                                    "addr": "node2_addr",
-                                    "link": "0",
-                                    "type": "FQDN",
-                                }
-                            ],
-                        },
-                    ],
-                    "links_options": {},
-                    "quorum_options": {},
-                    "quorum_device": None,
-                }
-            )
-            + "\n"
-        )
-
-    @staticmethod
-    def fixture_cmd_output():
-        return dedent(
-            """\
-            pcs cluster setup cluster_name \\
-              node1 addr=node1_addr \\
-              node2 addr=node2_addr \\
-              transport \\
-              knet \\
-                  ip_version=ipv6 \\
-                crypto \\
-                  cipher=aes256 \\
-                  hash=sha256
-            """
-        )
 
     def setUp(self):
         self.corosync_conf_file = get_tmp_file(
@@ -114,25 +121,25 @@ class ClusterConfigMixin(AssertPcsMixin):
     def test_default_output(self):
         self.assert_pcs_success(
             self.command.split(),
-            stdout_full=self.fixture_text_output(),
+            stdout_full=fixture_text_output(),
         )
 
     def test_text_output(self):
         self.assert_pcs_success(
             (self.command + " --output-format=text").split(),
-            stdout_full=self.fixture_text_output(),
+            stdout_full=fixture_text_output(),
         )
 
     def test_json_output(self):
         self.assert_pcs_success(
             (self.command + " --output-format=json").split(),
-            stdout_full=self.fixture_json_output(),
+            stdout_full=fixture_json_output(),
         )
 
     def test_cmd_output(self):
         self.assert_pcs_success(
             (self.command + " --output-format=cmd").split(),
-            stdout_full=self.fixture_cmd_output(),
+            stdout_full=fixture_cmd_output(),
         )
 
     def test_output_format_unsupported_value(self):
@@ -161,3 +168,46 @@ class ClusterConfig(ClusterConfigMixin, TestCase):
 
 class ClusterConfigShow(ClusterConfigMixin, TestCase):
     command = "cluster config show"
+
+
+class ClusterConfigNoUuid(AssertPcsMixin, TestCase):
+    # pylint: disable=invalid-name
+    def setUp(self):
+        self.corosync_conf_file_no_uuid = get_tmp_file(
+            "tier1_cluster_config_show_corosync_no_uuid.conf"
+        )
+        self.pcs_runner = PcsRunner(
+            cib_file=None,
+            corosync_conf_opt=self.corosync_conf_file_no_uuid.name,
+        )
+        write_data_to_tmpfile(
+            fixture_corosync_conf_minimal(no_cluster_uuid=True),
+            self.corosync_conf_file_no_uuid,
+        )
+
+    def tearDown(self):
+        self.corosync_conf_file_no_uuid.close()
+
+    def test_default_output(self):
+        self.assert_pcs_success(
+            ["cluster", "config", "show"],
+            stdout_full=fixture_text_output(no_cluster_uuid=True),
+        )
+
+    def test_text_output(self):
+        self.assert_pcs_success(
+            ["cluster", "config", "show", "--output-format=text"],
+            stdout_full=fixture_text_output(no_cluster_uuid=True),
+        )
+
+    def test_json_output_no_uuid(self):
+        self.assert_pcs_success(
+            ["cluster", "config", "show", "--output-format=json"],
+            stdout_full=fixture_json_output(no_cluster_uuid=True),
+        )
+
+    def test_cmd_output_no_uuid(self):
+        self.assert_pcs_success(
+            ["cluster", "config", "show", "--output-format=cmd"],
+            stdout_full=fixture_cmd_output(no_cluster_uuid=True),
+        )